requirements.txt
llama-index
numpy
python-dotenv
requests>=2.31.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
pandas>=2.2.0
plotly>=5.19.0
networkx>=3.2.1
jinja2>=3.1.0
tqdm>=4.66.1
